pyproject.toml
[tool.poetry]
name = "devchannel"
version = "0.1.0"
description = ""
authors = ["Marcus Pereira <marcus@negros.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "2.1.2"
Flask-SQLAlchemy = "2.5.1"
Flask-Migrate = "3.1.0"
dynaconf = "3.1.9"
factory-boy = "3.2.1"

[tool.poetry.dev-dependencies]
python-dotenv = "0.20.0"
pytest = "7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
